Move files to test folder
