+++ test/spec/no-force-option.spec.ts
@@ -1,5 +1,5 @@
-import { runRuleTester, wrapInTest } from '../../utils/rule-tester';
-import rule from '../no-force-option';
+import { runRuleTester, wrapInTest } from '../utils/rule-tester';
+import rule from '../../src/rules/no-force-option';
 
 const invalid = (code: string) => ({
   code: wrapInTest(code),
