+++ test/spec/no-wait-for-timeout.spec.ts
@@ -1,5 +1,5 @@
-import { runRuleTester } from '../../utils/rule-tester';
-import rule from '../no-wait-for-timeout';
+import { runRuleTester } from '../utils/rule-tester';
+import rule from '../../src/rules/no-wait-for-timeout';
 
 const invalid = (code: string, output: string) => ({
   code,
